Time: 15 ms (10.55%), Space: 8.4 MB (85.12%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long mySqrt(long long x) {
+    int mySqrt(long long x) {
         if(x==1)return 1;
  for(long long i=0;i<x+1;i++)
  {
